requirements.txt
astroid==2.15.3
beautifulsoup4==4.11.1
black==23.1.0
boto3
celery==4.*
crispy-bootstrap4
Django==4.1.1
django-crispy-forms
django-recaptcha
django-storages
dj-database-url==1.0.0
factory-boy==3.2.1
Faker==18.6.0
gevent
gunicorn
lxml==4.9.1
Pillow==8.4.0
psycopg2-binary==2.9.3
pylint==2.17.2
pytest-django==4.5.2
redis
requests==2.26.0
stripe==2.63.0
wagtail==4.2.1
wagtail-factories==4.0.0
whitenoise==6.2.0
